+++ Domain.Tests/Model/Signal.Tests.cs
@@ -1,6 +1,9 @@
 using System;
 using N17Solutions.Semaphore.Domain.Model;
 using N17Solutions.Semaphore.ServiceContract;
+using N17Solutions.Semaphore.ServiceContract.Signals;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using Shouldly;
 using Xunit;
 
@@ -88,6 +91,131 @@ namespace N17Solutions.Semaphore.Domain.Tests.Model
             
             // Assert
             result.IsEncrypted.ShouldBe(expected);
+        }
+
+        [Fact]
+        public void Should_Map_String_Value_To_Write_Model()
+        {
+            // Arrange
+            var domainModel = new Signal
+            {
+                Id = 1,
+                Name = "Test Signal",
+                Value = "Test Value",
+                ValueType = typeof(string).FullName,
+                IsBaseType = true,
+                Tags = "Test,Tags",
+                DateCreated = DateTime.Now,
+                DateLastUpdated = DateTime.Now
+            };
+            
+            // Act
+            var result = domainModel.ToWriteModel();
+            
+            // Assert
+            result.Name.ShouldBe(domainModel.Name);
+            result.Tags.ShouldBe(domainModel.Tags.Split(','));
+            result.Encrypted.ShouldBeFalse();
+            result.Value.ShouldBe(domainModel.Value);
+        }
+
+        [Fact]
+        public void Should_Map_Object_Value_To_WriteModel()
+        {
+            // Arrange
+            const string name = "Test Name";
+            
+            var domainModel = new Signal
+            {
+                Id = 1,
+                Name = "Test Signal",
+                Value = JsonConvert.SerializeObject(new SignalWriteModel { Name = name }),
+                ValueType = typeof(SignalWriteModel).AssemblyQualifiedName,
+                IsBaseType = false,
+                Tags = "Test,Tags",
+                DateCreated = DateTime.Now,
+                DateLastUpdated = DateTime.Now
+            };
+            
+            // Act
+            var result = domainModel.ToWriteModel();
+            
+            // Assert
+            result.Name.ShouldBe(domainModel.Name);
+            result.Tags.ShouldBe(domainModel.Tags.Split(','));
+            result.Encrypted.ShouldBeFalse();
+            result.Value.ShouldBeOfType<SignalWriteModel>();
+            ((SignalWriteModel)result.Value).Name.ShouldBe(name);
+        }
+
+        [Fact]
+        public void Should_Populate_From_Object_WriteModel()
+        {
+            // Arrange
+            var writeModel = new SignalWriteModel
+            {
+                Name = "Test Name",
+                Tags = new[]{"Test", "Tags"},
+                Encrypted = false,
+                Value = new SignalWriteModel { Name = "Child" }
+            };
+            var domainModel = new Signal();
+            
+            // Act
+            domainModel.PopulateFromWriteModel(writeModel);
+            
+            // Assert
+            domainModel.Name.ShouldBe(writeModel.Name);
+            domainModel.Tags.ShouldBe(string.Join(",", writeModel.Tags));
+            domainModel.Value.ShouldBe(JsonConvert.SerializeObject(writeModel.Value));
+        }
+
+        [Fact]
+        public void Should_Populate_From_Base_WriteModel()
+        {
+            // Arrange
+            var writeModel = new SignalWriteModel
+            {
+                Name = "Test Name",
+                Tags = new[]{"Test", "Tags"},
+                Encrypted = false,
+                Value = "Test Value"
+            };
+            var domainModel = new Signal();
+            
+            // Act
+            domainModel.PopulateFromWriteModel(writeModel);
+            
+            // Assert
+            domainModel.Name.ShouldBe(writeModel.Name);
+            domainModel.Tags.ShouldBe(string.Join(",", writeModel.Tags));
+            domainModel.Value.ShouldBe(writeModel.Value);
+        }
+
+        [Theory]
+        [InlineData(Constants.EncryptedTag, true)]
+        [InlineData("Tag", false)]
+        [InlineData(null, false)]
+        public void Should_Denote_If_IsEncrypted(string tag, bool expected)
+        {
+            // Arrange
+            var domainModel = new Signal
+            {
+                Id = 1,
+                Name = "Test Signal",
+                Value = "Test",
+                ValueType = typeof(string).FullName,
+                IsBaseType = true,
+                Tags = tag,
+                DateCreated = DateTime.Now,
+                DateLastUpdated = DateTime.Now
+            };
+            
+            // Act
+            var result = domainModel.IsEncrypted();
+            
+            // Assert
+            result.ShouldBe(expected);
         }
     }
 }