+++ Handlers.Tests/Signals/GetSignalByNameRequestHandler.Tests.cs
@@ -12,6 +12,7 @@ using N17Solutions.Semaphore.Handlers.Signals;
 using N17Solutions.Semaphore.Requests.Security;
 using N17Solutions.Semaphore.Requests.Signals;
 using N17Solutions.Semaphore.ServiceContract;
+using N17Solutions.Semaphore.ServiceContract.Extensions;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Shouldly;
@@ -107,7 +108,7 @@ namespace N17Solutions.Semaphore.Handlers.Tests.Signals
             
             // Assert
             result.ShouldNotBeNull();
-            result.Value.ShouldBeOfType<JObject>();
+            result.Value.ShouldBeOfType<TestObject>();
         }
         
         [Fact]