+++ Handlers/InternalsPointer.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly:InternalsVisibleTo("N17Solutions.Semaphore.Handlers.Tests")]