+++ Handlers/Signals/GetSignalByNameAndTagRequestHandler.cs
@@ -9,6 +9,7 @@ using N17Solutions.Semaphore.Domain.Model;
 using N17Solutions.Semaphore.Requests.Security;
 using N17Solutions.Semaphore.Requests.Signals;
 using N17Solutions.Semaphore.Responses.Signals;
+using N17Solutions.Semaphore.ServiceContract.Signals;
 using Newtonsoft.Json;
 
 // ReSharper disable InvertIf
@@ -48,9 +49,7 @@ namespace N17Solutions.Semaphore.Handlers.Signals
                 else if (result.IsEncrypted)
                     return result;
                 
-                var valueType = Type.GetType(result.ValueType);
-                if (valueType != null)
-                    result.Value = result.IsBaseType ? Convert.ChangeType(result.Value, valueType) : JsonConvert.DeserializeObject(result.Value.ToString());
+                result.Value = ValueResolver.Resolve(result.Value, result.ValueType, result.IsBaseType);
             }
 
             return result;