+++ Requests/Signals/CreateSignalRequest.cs
@@ -1,30 +1,9 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using MediatR;
+using N17Solutions.Semaphore.ServiceContract.Signals;
 
 namespace N17Solutions.Semaphore.Requests.Signals
 {
-    public class CreateSignalRequest : IRequest<Guid>
-    {
-        /// <summary>
-        /// The Name of this Signal.
-        /// </summary>
-        public string Name { get; set; }
-        
-        /// <summary>
-        /// The value to set on this Signal.
-        /// </summary>
-        public object Value { get; set; }
-
-        /// <summary>
-        /// A collection of tags to add to this signal.
-        /// </summary>
-        public IEnumerable<string> Tags { get; set; } = Enumerable.Empty<string>();
-
-        /// <summary>
-        /// Whether to Encrypt the value
-        /// </summary>
-        public bool Encrypted { get; set; }
-    }
+    public class CreateSignalRequest : SignalWriteModel, IRequest<Guid>
+    { }
 }