+++ ServiceContract/Extensions/ObjectExtensions.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace N17Solutions.Semaphore.Handlers.Extensions
+namespace N17Solutions.Semaphore.ServiceContract.Extensions
 {
     public static class ObjectExtensions
     {
@@ -8,9 +8,14 @@ namespace N17Solutions.Semaphore.Handlers.Extensions
         
         public static string GetSignalValueType(this object target)
         {
-            return target.IsBaseType()
-                ? target.GetType().FullName 
-                : new object().GetType().FullName;
+            try
+            {
+                return target.GetType().AssemblyQualifiedName;
+            }
+            catch
+            {
+                return new object().GetType().FullName;
+            }
         }
 
         public static bool IsBaseType(this object target)