+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+#### 1.1 - 29.10.2018
+
+  Signal Patching
+  
+    * Add Signal Patch Controller
+    * Add Signal Patch Requests
+    * Add Signal Patch Handlers
+
 #### 1.0 - 16.09.2018
 
   Initial Commit